feat: upgrade to EDC 0.6 and port the service patterns from embedded matchmaking.
