+++ .gitignore
@@ -35,7 +35,6 @@ __pycache__
 .classpath
 .project
 .settings/
-*.bak
 *cs-cleanup.xml
 *cs-formatter.xml
 .checkstyle
